Created a different branch for issue #5
